requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
databricks-sql-connector